+++ src/ifm3d_node.cpp
@@ -32,6 +32,10 @@
 #include <ifm3d/camera.h>
 #include <ifm3d/fg.h>
 #include <ifm3d/image.h>
+#include <ifm3d/Config.h>
+#include <ifm3d/Dump.h>
+#include <ifm3d/Extrinsics.h>
+#include <ifm3d/Trigger.h>
 
 namespace enc = sensor_msgs::image_encodings;
 
@@ -88,7 +92,29 @@ public:
     this->uvec_pub_ =
       nh.advertise<sensor_msgs::Image>("unit_vectors", 1, true);
 
-    //this->extrinsics_pub_ = nh.advertise<ifm3d::Extrinsics>("extrinsics", 1);
+    this->extrinsics_pub_ = nh.advertise<ifm3d::Extrinsics>("extrinsics", 1);
+
+    //---------------------
+    // Advertised Services
+    //---------------------
+    this->dump_srv_ =
+      nh.advertiseService<ifm3d::Dump::Request, ifm3d::Dump::Response>
+      ("Dump", std::bind(&IFM3DNode::Dump, this,
+                         std::placeholders::_1,
+                         std::placeholders::_2));
+
+    this->config_srv_ =
+      nh.advertiseService<ifm3d::Config::Request, ifm3d::Config::Response>
+      ("Config", std::bind(&IFM3DNode::Config, this,
+                           std::placeholders::_1,
+                           std::placeholders::_2));
+
+    this->trigger_srv_ =
+      nh.advertiseService<ifm3d::Trigger::Request, ifm3d::Trigger::Response>
+      ("Trigger", std::bind(&IFM3DNode::Trigger, this,
+                            std::placeholders::_1,
+                            std::placeholders::_2));
+
   } // end: ctor
 
   /**
@@ -261,6 +287,7 @@ public:
         depth_with_confidence_img = distance_img.clone();
         amplitude_img = this->im_->AmplitudeImage();
         raw_amplitude_img = this->im_->RawAmplitudeImage();
+        extrinsics = this->im_->Extrinsics();
 
         lock.unlock();
 
@@ -344,20 +371,123 @@ public:
           //publish zeros otherwise
           depth_with_confidence_img = cv::Mat::zeros(confidence_img.rows,
                                             confidence_img.cols,
-                                            CV_8UC1);
+                                            CV_32FC1);
         }
         
         sensor_msgs::ImagePtr depth_with_confidence = cv_bridge::CvImage(optical_head,
-                                                      "mono16",
+                                                      depth_with_confidence_img.type() == CV_32FC1 ?
+                                                      enc::TYPE_32FC1 : enc::TYPE_16UC1,
                                                       depth_with_confidence_img).toImageMsg();
 
         this->depth_with_confidence_pub_.publish(depth_with_confidence);
         //
-        // XXX: Need to publish extrinsics
+        // publish extrinsics
         //
+        ifm3d::Extrinsics extrinsics_msg;
+        extrinsics_msg.header = optical_head;
+        try
+          {
+            extrinsics_msg.tx = extrinsics.at(0);
+            extrinsics_msg.ty = extrinsics.at(1);
+            extrinsics_msg.tz = extrinsics.at(2);
+            extrinsics_msg.rot_x = extrinsics.at(3);
+            extrinsics_msg.rot_y = extrinsics.at(4);
+            extrinsics_msg.rot_z = extrinsics.at(5);
+          }
+        catch (const std::out_of_range& ex)
+          {
+            ROS_WARN("out-of-range error fetching extrinsics");
+          }
+        this->extrinsics_pub_.publish(extrinsics_msg);
 
       } // end: while(ros::ok()) {...}
   } // end: Run()
+
+  /**
+   * Implements the `Dump' service
+   *
+   * The `Dump' service will dump the current camera configuration to a JSON
+   * string. This JSON string is suitable for editing and using to reconfigure
+   * the camera via the `Config' service.
+   */
+  bool Dump(ifm3d::Dump::Request &req,
+            ifm3d::Dump::Response &res)
+  {
+    std::lock_guard<std::mutex> lock(this->mutex_);
+    res.status = 0;
+
+    try
+      {
+        res.config = this->cam_->ToJSONStr();
+      }
+    catch (const ifm3d::error_t& ex)
+      {
+        res.status = ex.code();
+      }
+
+    return true;
+  }
+
+  /**
+   * Implements the `Config' service.
+   *
+   * The `Config' service will read the input JSON configuration data and
+   * mutate the camera's settings to match that of the configuration
+   * described by the JSON file. Syntactically, the JSON should look like the
+   * JSON that is produced by `Dump'. However, you need not specify every
+   * parameter. You can specify only the parameters you wish to change with the
+   * only caveat being that you need to specify the parameter as fully
+   * qualified from the top-level root of the JSON tree.
+   */
+  bool Config(ifm3d::Config::Request &req,
+              ifm3d::Config::Response &res)
+  {
+    std::lock_guard<std::mutex> lock(this->mutex_);
+    res.status = 0;
+    res.msg = "OK";
+
+    try
+      {
+        this->cam_->FromJSONStr(req.json);
+      }
+    catch (const ifm3d::error_t& ex)
+      {
+        res.status = ex.code();
+        res.msg = ex.what();
+      }
+    catch (const std::exception& std_ex)
+      {
+        res.status = -1;
+        res.msg = std_ex.what();
+      }
+
+    return true;
+  }
+
+  /**
+   * Implements the `Trigger' service.
+   *
+   * For cameras whose active application is set to software triggering as
+   * opposed to free-running, this service send the trigger for image
+   * acquisition to the camera.
+   */
+  bool Trigger(ifm3d::Trigger::Request &req,
+               ifm3d::Trigger::Response &res)
+  {
+    std::lock_guard<std::mutex> lock(this->mutex_);
+    res.status = 0;
+
+    try
+      {
+        this->fg_->SWTrigger();
+      }
+    catch (const ifm3d::error_t& ex)
+      {
+        res.status = ex.code();
+      }
+
+    return true;
+  }
 
 private:
   std::string camera_ip_;